renamed webpack configs
