+++ internals/webpack/webpack.config.base.babel.js
